Replace game for videogame, modify tags
Refactor media type terminology from 'Game' to 'Videogame' across models, APIs, and settings for consistency and clarity. Also, replace default tag with 2 tags: media and the specific to the type of media. I had a bug when filtering the results when there was a null value, so I added an extra filter.